Update ImpressParser version to new Impress 2 version (Merge and update from mrelac's Impress branch)
